1. message id注册及存储修改, 减少int转Integer次数 2. Field定义优化, 新增PrimitiveUnsafeField, 减少primitive装包类型分配 3. 新增ThreadLocal cache, 缓存动态类型class, 减少getClass()调用次数 4. 移除bytebuddy支持

+++ kinbuffer/kinbuffer-core/src/main/java/org/kin/kinbuffer/io/ByteArrayOutput.java
@@ -68,7 +68,7 @@ public final class ByteArrayOutput implements Output{
     }
 
     @Override
-    public Output writeSInt64(int sl) {
+    public Output writeSInt64(long sl) {
         VarIntUtils.writeRawVarInt64(array, sl, true);
         return this;
     }
